+++ .azure/config
@@ -3,5 +3,5 @@ group = IoT_Hub_Resource_Group
 sku = F1
 appserviceplan = ASP-IoTHubResourceGroup-8260
 location = southcentralus
-web = wzdc-rest-api
+web = wzdc-rest-api-dev
