+++ startup.sh
@@ -1,0 +1,1 @@
+gunicorn -w 4 -k uvicorn.workers.UvicornWorker main:app
